Move a couple of Akka Streams examples to their own files
Might make it easier to follow?
