+++ akka-docs/src/main/paradox/stream/operators/Source-or-Flow/scan.md
@@ -34,7 +34,7 @@ Note that the `zero` value must be immutable.
 ## Examples
 
 Scala
-:  @@snip [SourceOrFlow.scala](/akka-docs/src/test/scala/docs/stream/operators/SourceOrFlow.scala) { #scan }
+:  @@snip [SourceOrFlow.scala](/akka-docs/src/test/scala/docs/stream/operators/sourceorflow/Scan.scala) { #scan }
 
 Java
 :  @@snip [SourceOrFlow.java](/akka-docs/src/test/java/jdocs/stream/operators/SourceOrFlow.java) { #scan }
